Flytt barnehagestartDato til delt utils-pakke
Oppretter packages/utils/src/barnehagestartUtils.ts med
barnehagestartDato-funksjonen, eksportert frå @navikt/fp-utils.

Eksporterer isIkkeUtfyltTypeBarn frå @navikt/fp-types offentleg API.

Oppdaterer UttaksplanSteg.tsx til å bruke den delte funksjonen i
staden for lokal implementasjon.
