+++ apps/foreldrepengesoknad/src/steps/uttaksplan/UttaksplanSteg.tsx
@@ -3,7 +3,6 @@ import { useQuery } from '@tanstack/react-query';
 import { useAnnenPartVedtakOptions, useStønadsKontoerOptions } from 'api/queries';
 import { ContextDataType, useContextGetData, useContextSaveData } from 'appData/FpDataContext';
 import { useStepConfig } from 'appData/useStepConfig';
-import dayjs from 'dayjs';
 import { ReactNode, useRef, useState } from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { isAnnenForelderOppgitt } from 'types/AnnenForelder';
@@ -14,19 +13,14 @@ import { getErSøkerFarEllerMedmor, getKjønnFromFnr, getNavnPåForeldre } from 
 
 import { Alert, BodyLong, Tabs } from '@navikt/ds-react';
 
-import { ISO_DATE_FORMAT } from '@navikt/fp-constants';
 import { loggUmamiEvent } from '@navikt/fp-observability';
 import {
-    Barn,
     FpPersonopplysningerDto_fpoversikt,
     FpSak_fpoversikt,
     RettighetType_fpoversikt,
-    isAdoptertBarn,
-    isFødtBarn,
 } from '@navikt/fp-types';
-import { isIkkeUtfyltTypeBarn } from '@navikt/fp-types/src/Barn';
 import { SkjemaRotLayout, Step } from '@navikt/fp-ui';
-import { Uttaksperioden, getFamiliehendelsedato } from '@navikt/fp-utils';
+import { Uttaksperioden, barnehagestartDato, getFamiliehendelsedato } from '@navikt/fp-utils';
 import {
     FjernAltIUttaksplanModal,
     HvaErMulig,
@@ -243,54 +237,6 @@ const utledRettighet = (erAleneOmOmsorg: boolean, erDeltUttak: boolean): Rettigh
         return 'BEGGE_RETT';
     }
     return 'BARE_SØKER_RETT';
-};
-
-// TODO (TOR) Flytt denne til felleskode - er lik funksjon i planlegger
-const barnehagestartDato = (barnet: Barn) => {
-    if (isAdoptertBarn(barnet) || isIkkeUtfyltTypeBarn(barnet)) {
-        return undefined;
-    }
-
-    const dato = isFødtBarn(barnet) ? barnet.fødselsdatoer[0]! : barnet.termindato;
-
-    if (dayjs(dato).month() < 8) {
-        const newLocal = dayjs(dato).month(7).add(1, 'year').endOf('month').format(ISO_DATE_FORMAT);
-        return getUttaksdagTilOgMedDato(newLocal);
-    }
-    if (dayjs(dato).month() >= 8 && dayjs(dato).month() < 11) {
-        return getUttaksdagTilOgMedDato(dayjs(dato).add(1, 'year').endOf('month').format(ISO_DATE_FORMAT));
-    }
-    return getUttaksdagTilOgMedDato(
-        dayjs(dato)
-            .startOf('year')
-            .add(2, 'year')
-            .add(7, 'months')
-            .endOf('week')
-            .endOf('month')
-            .format(ISO_DATE_FORMAT),
-    );
-};
-
-/**
- * Sjekker om dato er en ukedag, dersom ikke finner den foregående fredag.
- * Tar hensyn til stilling av klokken ved å gjøre om klokka til kl 12 før antall timer trekkes fra.
- * @param dato
- */
-const getUttaksdagTilOgMedDato = (dato: string): string => {
-    const d = dayjs(dato).toDate();
-    const newDate = dato ? new Date(d.getFullYear(), d.getMonth(), d.getDate(), 12) : dato;
-    switch (getUkedag(dato)) {
-        case 6:
-            return dayjs.utc(newDate).subtract(24, 'hours').format(ISO_DATE_FORMAT);
-        case 7:
-            return dayjs.utc(newDate).subtract(48, 'hours').format(ISO_DATE_FORMAT);
-        default:
-            return dato;
-    }
-};
-
-const getUkedag = (dato: string): number => {
-    return dayjs(dato).isoWeekday();
 };
 
 const loggExpansionCardOpen = (tittel: string) => (open: boolean) => {
