+++ packages/utils/index.ts
@@ -84,3 +84,4 @@ export { Uttaksperioden } from './src/uttak/Uttaksperioden';
 export { Tidsperioden } from './src/Tidsperioden';
 export * from './src/cookieUtils';
 export { periodFormat } from './src/periodUtils';
+export { barnehagestartDato } from './src/barnehagestartUtils';
