+++ packages/utils/src/barnehagestartUtils.ts
@@ -1,0 +1,28 @@
+import dayjs from 'dayjs';
+
+import { ISO_DATE_FORMAT } from '@navikt/fp-constants';
+import { Barn, isAdoptertBarn, isFødtBarn, isIkkeUtfyltTypeBarn } from '@navikt/fp-types';
+
+import { Uttaksdagen } from './uttak/Uttaksdagen';
+
+export const barnehagestartDato = (barnet: Barn): string | undefined => {
+    if (isAdoptertBarn(barnet) || isIkkeUtfyltTypeBarn(barnet)) {
+        return undefined;
+    }
+
+    const dato = isFødtBarn(barnet) ? barnet.fødselsdatoer[0]! : barnet.termindato;
+
+    if (dayjs(dato).month() < 8) {
+        return Uttaksdagen.denneEllerForrige(
+            dayjs(dato).month(7).add(1, 'year').endOf('month').format(ISO_DATE_FORMAT),
+        ).getDato();
+    }
+    if (dayjs(dato).month() >= 8 && dayjs(dato).month() < 11) {
+        return Uttaksdagen.denneEllerForrige(
+            dayjs(dato).add(1, 'year').endOf('month').format(ISO_DATE_FORMAT),
+        ).getDato();
+    }
+    return Uttaksdagen.denneEllerForrige(
+        dayjs(dato).startOf('year').add(2, 'year').add(7, 'months').endOf('week').endOf('month').format(ISO_DATE_FORMAT),
+    ).getDato();
+};
